requirements.txt
python-dotenv
numpy
matplotlib
aiohttp
requests